Re-add Stats Through Skills
Contrary to a claim on the github, it doesn't appear to be possible to
recreate STS' functionality through the enchantments system. At least
not *sanely* -- you *might* be able to hack together something using
enchantments and a heaping mountain of EOCs, but it would be a very bad
idea.

I'm still leaving it marked as obsoleted, though, because ultimately I'd
like to see a proper mechanism for naturalistic stat improvements added
to the core game.