Ensure void spider rewards will always be spawned
Previously, it was possible to miss out on the void spider rewards if
you killed the void weaver *too quickly*, which was a real issue that
multiple people ran into.

Having to know *a priori* that you need to leave the thing alive long
enough for it to warp the terrain a couple of times in order to get
anything good out of the fight is bad game design, so now if you kill
the void weaver and either the egg room or the corpse room *haven't* yet
been warped in, they will be after 10 seconds.